[RUM] extract resource utils + fix resource size
