+++ src/rum/rum.ts
@@ -1,8 +1,9 @@
 import { Configuration } from '../core/configuration'
 import { ErrorContext, ErrorMessage, ErrorObservable, HttpContext } from '../core/errorCollection'
-import { addMonitoringMessage, monitor } from '../core/internalMonitoring'
+import { monitor } from '../core/internalMonitoring'
 import { Batch, HttpRequest } from '../core/transport'
 import { generateUUID, msToNs, ResourceKind, withSnakeCaseKeys } from '../core/utils'
+import { computePerformanceResourceDetails, computeResourceKind, computeSize, isValidResource } from './resourceUtils'
 import { RumSession } from './rumSession'
 
 declare global {
@@ -29,7 +30,7 @@ interface PerformanceResourceDetailsElement {
   start: number
 }
 
-interface PerformanceResourceDetails {
+export interface PerformanceResourceDetails {
   redirect?: PerformanceResourceDetailsElement
   dns: PerformanceResourceDetailsElement
   connect: PerformanceResourceDetailsElement
@@ -48,7 +49,7 @@ export interface RumResourceEvent {
     url: string
   }
   network: {
-    bytesWritten: number
+    bytesWritten?: number
   }
   resource: {
     kind: ResourceKind
@@ -94,25 +95,6 @@ export interface RumErrorEvent {
 }
 
 export type RumEvent = RumErrorEvent | RumPerformanceScreenEvent | RumResourceEvent
-
-const RESOURCE_TYPES: Array<[ResourceKind, (initiatorType: string, path: string) => boolean]> = [
-  [ResourceKind.XHR, (initiatorType: string) => 'xmlhttprequest' === initiatorType],
-  [ResourceKind.FETCH, (initiatorType: string) => 'fetch' === initiatorType],
-  [ResourceKind.BEACON, (initiatorType: string) => 'beacon' === initiatorType],
-  [ResourceKind.CSS, (_: string, path: string) => path.match(/\.css$/i) !== null],
-  [ResourceKind.JS, (_: string, path: string) => path.match(/\.js$/i) !== null],
-  [
-    ResourceKind.IMAGE,
-    (initiatorType: string, path: string) =>
-      ['image', 'img', 'icon'].includes(initiatorType) || path.match(/\.(gif|jpg|jpeg|tiff|png|svg)$/i) !== null,
-  ],
-  [ResourceKind.FONT, (_: string, path: string) => path.match(/\.(woff|eot|woff2|ttf)$/i) !== null],
-  [
-    ResourceKind.MEDIA,
-    (initiatorType: string, path: string) =>
-      ['audio', 'video'].includes(initiatorType) || path.match(/\.(mp3|mp4)$/i) !== null,
-  ],
-]
 
 export let pageViewId: string
 let activeLocation: Location
@@ -242,7 +224,7 @@ export function handleResourceEntry(
   entry: PerformanceResourceTiming,
   addRumEvent: (event: RumEvent) => void
 ) {
-  if (!entry.name || isBrowserAgentRequest(entry.name, configuration)) {
+  if (!isValidResource(entry.name, configuration)) {
     return
   }
   const resourceKind = computeResourceKind(entry)
@@ -257,7 +239,7 @@ export function handleResourceEntry(
       url: entry.name,
     },
     network: {
-      bytesWritten: entry.decodedBodySize,
+      bytesWritten: computeSize(entry),
     },
     resource: {
       kind: resourceKind,
@@ -268,62 +250,6 @@ export function handleResourceEntry(
         }
       : undefined,
   })
-}
-
-function isBrowserAgentRequest(url: string, configuration: Configuration) {
-  return (
-    url.startsWith(configuration.logsEndpoint) ||
-    url.startsWith(configuration.rumEndpoint) ||
-    (configuration.internalMonitoringEndpoint && url.startsWith(configuration.internalMonitoringEndpoint))
-  )
-}
-
-function computePerformanceResourceDetails(entry: PerformanceResourceTiming): PerformanceResourceDetails | undefined {
-  if (hasTimingAllowedAttributes(entry)) {
-    return {
-      connect: { duration: msToNs(entry.connectEnd - entry.connectStart), start: msToNs(entry.connectStart) },
-      dns: {
-        duration: msToNs(entry.domainLookupEnd - entry.domainLookupStart),
-        start: msToNs(entry.domainLookupStart),
-      },
-      download: { duration: msToNs(entry.responseEnd - entry.responseStart), start: msToNs(entry.responseStart) },
-      firstByte: { duration: msToNs(entry.responseStart - entry.requestStart), start: msToNs(entry.requestStart) },
-      redirect:
-        entry.redirectStart > 0
-          ? { duration: msToNs(entry.redirectEnd - entry.redirectStart), start: msToNs(entry.redirectStart) }
-          : undefined,
-      ssl:
-        entry.secureConnectionStart > 0
-          ? {
-              duration: msToNs(entry.connectEnd - entry.secureConnectionStart),
-              start: msToNs(entry.secureConnectionStart),
-            }
-          : undefined,
-    }
-  }
-  return undefined
-}
-
-function hasTimingAllowedAttributes(timing: PerformanceResourceTiming) {
-  return timing.responseStart > 0
-}
-
-function computeResourceKind(timing: PerformanceResourceTiming) {
-  let url: URL | undefined
-  try {
-    url = new URL(timing.name)
-  } catch (e) {
-    addMonitoringMessage(`Failed to construct URL for "${timing.name}"`)
-  }
-  if (url !== undefined) {
-    const path = url.pathname
-    for (const [type, isType] of RESOURCE_TYPES) {
-      if (isType(timing.initiatorType, path)) {
-        return type
-      }
-    }
-  }
-  return ResourceKind.OTHER
 }
 
 export function handleNavigationEntry(entry: PerformanceNavigationTiming, addRumEvent: (event: RumEvent) => void) {
